+++ core-strato/milena/CHANGELOG.md
@@ -1,3 +1,13 @@
+0.5.4.0
+
+* commitOffset and fetchOffset to the API (https://github.com/adamflott/milena/pull/7)
+* Add Delete Topic support
+* Make tests pass on multiple invocations
+
+0.5.3.0
+
+* Add Create Topic support (https://github.com/adamflott/milena/pull/6). Thanks matth!
+
 0.5.2.3
 
 * Bump network upper bounds